switch from ECC to RSA and fix compiler bugs
ECC is new and sexy, but there's some ambiguity about which curves are
trustworthy, and it's not clear to me which standards have mature
support among implementations for both signatures and encryption.

RSA, on the other hand is well established and has stood the test of
time pretty well, provided you're using large key sizes.  It's slow,
but we won't be using it on any fast paths anyway.  The keys are big,
but we won't be transmitting them often.  Yay, RSA!